PrivCount can count parameters for traffic modeling
PrivCount takes in an HMM traffic model, and counts some items
needed to understand how well the data fits those models. The
output of this counting will be used to tune the model for future
rounds.

We still need to change the way sigma is assigned to the counters
used for fitting data to the model.

refs #19